backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "clerk-backend-api>=2.2.0",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1", # Changed from 0.25.0 to 0.28.1
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-text-splitters>=0.3.8",
    "markdown>=3.8.2",
    "odfpy>=1.4.1",
    "openai>=1.79.0",
    "pdf2image>=1.17.0",
    "pinecone>=7.3.0",
    "pymongo>=4.13.2",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.0",
    "python-pptx>=1.0.2",
    "qdrant-client>=1.15.0",
    "sqlalchemy>=2.0.41",
    "svix>=1.65.0",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.34.2",
]